+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/jenkinsci/analysis-model/compare/analysis-model-5.0.0...master)
 
+### Fixed
+- [JENKINS-57365](https://issues.jenkins-ci.org/browse/JENKINS-57365),
+[PR#174](https://github.com/jenkinsci/analysis-model/pull/174): 
+Fixes false positives for MsBuild if e.g. a build project name contains the string info/note/warning.
+
 ## [5.0.1](https://github.com/jenkinsci/analysis-model/compare/analysis-model-5.0.0...analysis-model-5.0.1) - 2019-5-12
 
 - [JENKINS-57379](https://issues.jenkins-ci.org/browse/JENKINS-57379),